+++ example/example/urls.py
@@ -1,13 +1,13 @@
-from django.conf.urls import include, url
+from django.urls import include, re_path
 from django.contrib import admin
 from wagtail.admin import urls as wagtailadmin_urls
 from wagtail.documents import urls as wagtaildocs_urls
 from wagtail.core import urls as wagtail_urls
 
 urlpatterns = [
-    url(r'^admin/', admin.site.urls),
-    url(r'^cms/', include(wagtailadmin_urls)),
-    url(r'^documents/', include(wagtaildocs_urls)),
-    url(r'^pages/', include(wagtail_urls)),
-    url(r'', include(wagtail_urls)),
+    re_path(r'^admin/', admin.site.urls),
+    re_path(r'^cms/', include(wagtailadmin_urls)),
+    re_path(r'^documents/', include(wagtaildocs_urls)),
+    re_path(r'^pages/', include(wagtail_urls)),
+    re_path(r'', include(wagtail_urls)),
 ]