+++ src/wagtailthemes/middleware.py
@@ -1,21 +1,25 @@
 from django.core.exceptions import ImproperlyConfigured
-from django.utils.deprecation import MiddlewareMixin
 from wagtail.core.models import Site
 
 from wagtailthemes.models import ThemeSettings
 from wagtailthemes.thread import set_theme
 
 
-class ThemeMiddleware(MiddlewareMixin):
-    def process_request(self, request):
+class ThemeMiddleware:
+    def __init__(self, get_response):
+        self.get_response = get_response
+
+    def __call__(self, request):
         site = Site.find_for_request(request)
 
-        if not site:
-            raise ImproperlyConfigured(
-                "Site not found!")
+        if site is None:
+            raise ImproperlyConfigured("Site not found!")
 
         theme_settings = ThemeSettings.for_site(site)
         theme = theme_settings.theme
 
-        if theme:
+        if theme is not None:
             set_theme(theme)
+
+        response = self.get_response(request)
+        return response