+++ src/wagtailthemes/tests/urls.py
@@ -1,8 +1,8 @@
-from django.conf.urls import include, url
+from django.urls import include, re_path
 from wagtail.admin import urls as wagtailadmin_urls
 from wagtail.core import urls as wagtail_urls
 
 urlpatterns = [
-    url(r'^admin/', include(wagtailadmin_urls)),
-    url(r'', include(wagtail_urls)),
+    re_path(r'^admin/', include(wagtailadmin_urls)),
+    re_path(r'', include(wagtail_urls)),
 ]